fix attribute type encoding in Shader.jl

+++ WGLMakie/src/Lines.js
@@ -52,7 +52,7 @@ function lines_vertex_shader(uniforms, attributes, is_linesegments) {
             flat out vec2 f_extrusion;          // invalid / not needed
             flat out float f_linewidth;
             flat out vec4 f_pattern_overwrite;  // invalid / not needed
-            flat out int f_instance_id;
+            flat out uint f_instance_id;
             flat out ${color} f_color1;
             flat out ${color} f_color2;
             flat out float f_alpha_weight;
@@ -188,7 +188,6 @@ function lines_vertex_shader(uniforms, attributes, is_linesegments) {
                 // used to compute width sdf
                 f_linewidth = halfwidth;
 
-                // TODO: this line crashes when lineindex_start is of type uint, why?
                 f_instance_id = lineindex_start; // NOTE: this is correct, no need to multiple by 2
 
                 // we restart patterns for each segment
@@ -253,7 +252,7 @@ function lines_vertex_shader(uniforms, attributes, is_linesegments) {
             flat out vec2 f_extrusion;
             flat out float f_linewidth;
             flat out vec4 f_pattern_overwrite;
-            flat out int f_instance_id;
+            flat out uint f_instance_id;
             flat out ${color} f_color1;
             flat out ${color} f_color2;
             flat out float f_alpha_weight;
@@ -641,7 +640,6 @@ function lines_vertex_shader(uniforms, attributes, is_linesegments) {
                 // used to compute width sdf
                 f_linewidth = halfwidth;
 
-                // TODO: this line crashes when lineindex_start is of type uint, why?
                 f_instance_id = lineindex_start;
 
                 f_cumulative_length = lastlen_start;
@@ -751,7 +749,7 @@ function lines_fragment_shader(uniforms, attributes) {
     flat in ${color} f_color1;
     flat in ${color} f_color2;
     flat in float f_alpha_weight;
-    flat in int f_instance_id;
+    flat in uint f_instance_id;
     flat in float f_cumulative_length;
     flat in ivec2 f_capmode;
     flat in vec4 f_linepoints;
@@ -998,7 +996,7 @@ function lines_fragment_shader(uniforms, attributes) {
 
         if (picking) {
             if (color.a > 0.1) {
-                fragment_color = pack_int(object_id, uint(f_instance_id));
+                fragment_color = pack_int(object_id, f_instance_id);
             }
             return;
         }
